requirements.txt
asgiref==3.8.1
Django==5.1
django-ratelimit==4.1.0
djangorestframework==3.15.2
sqlparse==0.5.1
typing_extensions==4.12.2
